- moved generateID() to a shared util folder
