+++ packages/ui/src/utils/generateId.ts
@@ -1,0 +1,6 @@
+export const generateID = () => {
+  const url = URL.createObjectURL(new Blob());
+  const [id] = url.toString().split('/').reverse();
+  URL.revokeObjectURL(url);
+  return id.replace('blob:nodedata:', '');
+};
